Unix: Code refactoring with bugfix related to port closed/nil state
