+++ go.mod
@@ -4,5 +4,7 @@ require (
 	github.com/creack/goselect v0.0.0-20180501195510-58854f77ee8d
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/stretchr/testify v1.3.0
+	go.uber.org/atomic v1.3.2 // indirect
+	go.uber.org/multierr v1.1.0 // indirect
 	golang.org/x/sys v0.0.0-20190225065934-cc5685c2db12
 )
